requirements.txt
piper-tts>=1.0
playwright>=1.40
pyyaml>=6.0
